Update to include functionality to clear errors on 'Otherwise' destination element
